+++ EatMeApp/src/EatMeApp/Controllers/CommensalController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using EatMeApp.Models;
 using EatMeApp.Utilities;
+using Microsoft.AspNetCore.Authorization;
 
 // For more information on enabling Web API for empty projects, visit http://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -33,35 +34,37 @@ namespace EatMeApp.Controllers
         public Commensal Get(int id)
         {
             var token = Request.Headers["Authorization"].ToString();
-            if (Authorizer.HasAccess(token, _context))
+            if (!string.IsNullOrWhiteSpace(token))
             {
-                try
+                token = token.Substring(7);
+                if (Authorizer.HasAccess(token, _context))
                 {
+                    try
+                    {
 
-                    var commensal = _context.Commnesals.SingleOrDefault(x => x.Id == id);
+                        var commensal = _context.Commnesals.SingleOrDefault(x => x.Id == id);
 
-                    return commensal;
+                        return commensal;
 
-                }
-                catch (Exception ex)
-                {
+                    }
+                    catch (Exception ex)
+                    {
 
-                    throw ex;
+                        throw ex;
+                    }
                 }
             }
-            else
-            {
-                return null;
-            }
+            return null;
         }
 
         // POST api/values
         [HttpPost]
+        [AllowAnonymous]
         public void Post([FromBody]Commensal commensal)
         {
-            var token = Request.Headers["Authorization"].ToString();
-            if (Authorizer.HasAccess(token, _context))
-            {
+            //var token = Request.Headers["Authorization"].ToString();
+            //if (Authorizer.HasAccess(token, _context))
+            //{
                 try
                 {
                     var maxId = _context.Commnesals.Max(x => x.Id);
@@ -79,7 +82,7 @@ namespace EatMeApp.Controllers
 
                     throw ex;
                 }
-            }
+            //}
 
 
         }
@@ -89,34 +92,37 @@ namespace EatMeApp.Controllers
         public void Put(int id, [FromBody]Commensal commensal)
         {
             var token = Request.Headers["Authorization"].ToString();
-            if (Authorizer.HasAccess(token, _context))
+            if (!string.IsNullOrWhiteSpace(token))
             {
-                try
+                token = token.Substring(7);
+                if (Authorizer.HasAccess(token, _context))
                 {
-                    var com = _context.Commnesals.SingleOrDefault(x => x.Id == id);
-                    if (com != null)
+                    try
                     {
-                        com.FirstName = commensal.FirstName;
-                        com.LastName = commensal.LastName;
-                        com.Address = commensal.Address;
-                        com.Preferences = commensal.Preferences;
-                        com.EmailAddress = commensal.EmailAddress;
-                        com.IdentityCard = commensal.IdentityCard;
-                        com.Password = commensal.Password;
-                        com.PostalCode = commensal.PostalCode;
-                        com.Username = commensal.Username;
-                        com.Phone = commensal.Phone;
-                        _context.SaveChanges();
+                        var com = _context.Commnesals.SingleOrDefault(x => x.Id == id);
+                        if (com != null)
+                        {
+                            com.FirstName = commensal.FirstName;
+                            com.LastName = commensal.LastName;
+                            com.Address = commensal.Address;
+                            com.Preferences = commensal.Preferences;
+                            com.EmailAddress = commensal.EmailAddress;
+                            com.IdentityCard = commensal.IdentityCard;
+                            com.Password = commensal.Password;
+                            com.PostalCode = commensal.PostalCode;
+                            com.Username = commensal.Username;
+                            com.Phone = commensal.Phone;
+                            _context.SaveChanges();
+                        }
+
                     }
+                    catch (Exception ex)
+                    {
 
-                }
-                catch (Exception ex)
-                {
-
-                    throw ex;
+                        throw ex;
+                    }
                 }
             }
-
         }
 
         // DELETE api/values/5
@@ -124,22 +130,26 @@ namespace EatMeApp.Controllers
         public void Delete(int id)
         {
             var token = Request.Headers["Authorization"].ToString();
-            if (Authorizer.HasAccess(token, _context))
+            if (!string.IsNullOrWhiteSpace(token))
             {
-                try
+                token = token.Substring(7);
+                if (Authorizer.HasAccess(token, _context))
                 {
-                    var commensal = _context.Commnesals.SingleOrDefault(x => x.Id == id);
-                    if (commensal != null)
+                    try
                     {
-                        _context.Commnesals.Remove(commensal);
-                        _context.SaveChanges();
+                        var commensal = _context.Commnesals.SingleOrDefault(x => x.Id == id);
+                        if (commensal != null)
+                        {
+                            _context.Commnesals.Remove(commensal);
+                            _context.SaveChanges();
+                        }
+
                     }
+                    catch (Exception ex)
+                    {
 
-                }
-                catch (Exception ex)
-                {
-
-                    throw ex;
+                        throw ex;
+                    }
                 }
             }
         }
